requirements.txt
asgiref==3.3.4
Django==3.2.4
django-appconf==1.0.4
django-imagekit==4.0.2
pilkit==2.0
Pillow==8.2.0
python-slugify==5.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
